first draft of something useful with scopus
