+++ search/src/mapwisefox/search/parser/adapters/__init__.py
