+++ search/src/mapwisefox/search/parser/adapters/_springer.py
