+++ search/src/mapwisefox/search/parser/grammar.lark
@@ -3,7 +3,7 @@ start: expr
 
 
 output_spec_expr: "[" "->" output_target ":" compound_expr "]"
-!output_target: "query" | "filter" | "both"
+!output_target: "query" | "filter"
 attr_clause: "in" field_list
 field_list: field_name ("," field_name)*
